+++ Adapters/BidonAdapterStartIo/Models/StartIoAdUnitExtras.swift
@@ -1,0 +1,14 @@
+//
+//  StartIoAdUnitExtras.swift
+//  BidonAdapterStartIo
+//
+
+import Foundation
+
+struct StartIoAdUnitExtras: Codable {
+    var tagId: String
+}
+
+struct StartIoBiddingTokenExtras: Codable { }
+
+