+++ Adapters/BidonAdapterStartIo/Models/StartIoParameters.swift
@@ -1,0 +1,11 @@
+//
+//  StartIoParameters.swift
+//  BidonAdapterStartIo
+//
+
+
+import Foundation
+
+public struct StartIoParameters: Codable {
+    var appId: String
+}