+++ Sandbox/Sandbox/Initialization/InitializationViewModel.swift
@@ -27,6 +27,7 @@ import BidonAdapterMyTarget
 import BidonAdapterChartboost
 import BidonAdapterIronSource
 import BidonAdapterMoloco
+import BidonAdapterStartIo
 import BidonAdapterTaurusX
 
 
@@ -154,6 +155,7 @@ fileprivate extension Array where Element == Bidon.Adapter {
             InMobiDemandSourceAdapter(),
             AmazonDemandSourceAdapter(),
             GoogleAdManagerDemandSourceAdapter(),
+            StartIoDemandSourceAdapter(),
             MolocoDemandSourceAdapter(),
             TaurusXDemandSourceAdapter()
         ].sorted { $0.demandId < $1.demandId }