+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - BDN-1069 Fixes in BCA banners
 - BDN-1010 Adapter Versioning and Github Actions
 - BDN-1017 TaurusX implementation
+- BDN-1056 Start.io implementation
 
 ## Fixes
 - BDN-1053 Location waiting fix